feat: results table column header styles
