+++ packages/frontend/src/hooks/tableVisualization/useTableConfig.ts
@@ -118,8 +118,7 @@ const useTableConfig = (
                 itemsMap,
                 resultsData,
                 isColumnVisible,
-                getHeader,
-                getDefaultColumnLabel,
+                showTableNames,
             });
         }
     }, [
@@ -130,6 +129,7 @@ const useTableConfig = (
         isColumnVisible,
         getHeader,
         getDefaultColumnLabel,
+        showTableNames,
     ]);
 
     // Remove columProperties from map if the column has been removed from results
